rename : 디렉토리 구조 변경
* Sidebar / DocumentEditPage 디렉토리 생성
